Code refactored, one composable per file, one enum class per file.
